Create new_main.py
This is the changed main.py. We had to change the planet scale because, if not, Lagrange was displayed inside of Earth. Lagrange is textured with Neptune because I didn't know what image would be appropriate for it. Likewise, Asteroid is textured with Earth as a temporary placeholder. The satellite is there but it's orbiting the wrong plane, couldn't figure out how to change that because when I did it just left the solar system everytime, I look into that more.